PRS: Added linters and code coverage (#11)
