+++ .github/workflows/partsrelationshipservice.yml
@@ -1,0 +1,32 @@
+name: PRS
+
+on:
+  pull_request:
+    branches: [ main ]
+    paths:
+      - 'coreservices/partsrelationshipservice/**'
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Set up JDK 11
+        uses: actions/setup-java@v2
+        with:
+          java-version: '11'
+          distribution: 'adopt'
+
+      - name: Cache maven packages
+        uses: actions/cache@v2
+        with:
+          path: ~/.m2
+          key: ${{ runner.os }}-m2-${{ hashFiles('**/pom.xml') }}
+          restore-keys: ${{ runner.os }}-m2
+
+      - name: Build with Maven
+        run: |
+          cd coreservices/partsrelationshipservice
+          mvn --batch-mode --update-snapshots verify
