--- coreservices/partrelationshipservice/src/main/java/net/catenax/prs/PrsApplication.java
@@ -1,13 +1,0 @@
-package net.catenax.prs;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PrsApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PrsApplication.class, args);
-	}
-
-}
