+++ coreservices/partsrelationshipservice/ci/pmd-rules.xml
@@ -1,0 +1,46 @@
+<?xml version="1.0"?>
+<ruleset name="Java Rules"
+        xmlns="http://pmd.sourceforge.net/ruleset/2.0.0"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://pmd.sourceforge.net/ruleset/2.0.0 https://pmd.sourceforge.io/ruleset_2_0_0.xsd">
+    <description>
+            Rules for Java
+    </description>
+
+    <rule ref="category/java/bestpractices.xml">
+        <!-- Too many false positives, seems to be broken when using var -->
+        <exclude name="UnusedPrivateMethod"/>
+    </rule>
+    <rule ref="category/java/codestyle.xml">
+        <!-- Small methods are only bloated, big methods should raise eyebrows in code review -->
+        <exclude name="OnlyOneReturn"/>
+        <!-- Too many false positives -->
+        <exclude name="LongVariable"/>
+        <!-- Conflicts quite often with unnecessary constructor rule -->
+        <exclude name="AtLeastOneConstructor"/>
+        <!-- Clashes with the usage of var -->
+        <exclude name="UseDiamondOperator"/>
+    </rule>
+    <rule ref="category/java/design.xml">
+        <exclude name="LoosePackageCoupling"/>
+        <!-- There are too many false positives especially with streams -->
+        <exclude name="LawOfDemeter"/>
+    </rule>
+    <rule ref="category/java/design.xml/UseUtilityClass">
+        <properties>
+            <property name="violationSuppressXPath" value="//ClassOrInterfaceDeclaration/preceding-sibling::Annotation/MarkerAnnotation/Name[@Image='SpringBootApplication']" />
+        </properties>
+    </rule>
+    <rule ref="category/java/documentation.xml"></rule>
+    <rule ref="category/java/errorprone.xml"></rule>
+    <rule ref="category/java/multithreading.xml"></rule>
+    <rule ref="category/java/performance.xml"></rule>
+    <rule ref="category/java/security.xml"></rule>
+
+    <rule ref="category/java/documentation.xml/CommentSize">
+        <properties>
+            <property name="maxLines" value="15" />
+            <property name="maxLineLength" value="100" />
+        </properties>
+    </rule>
+</ruleset>
