+++ coreservices/partsrelationshipservice/lombok.config
@@ -1,0 +1,3 @@
+# Add @Generated annotation to generated code, so that it is skipped by JaCoCo
+lombok.addLombokGeneratedAnnotation = true
+
