+++ coreservices/partsrelationshipservice/src/main/java/net/catenax/prs/PrsApplication.java
@@ -1,0 +1,30 @@
+//
+// Copyright (c) 2021 Copyright Holder (Catena-X Consortium)
+//
+// See the AUTHORS file(s) distributed with this work for additional
+// information regarding authorship.
+//
+// See the LICENSE file(s) distributed with this work for
+// additional information regarding license terms.
+//
+package net.catenax.prs;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * Application entry point.
+ */
+@SpringBootApplication
+public class PrsApplication {
+
+    /**
+     * Entry point.
+     *
+     * @param args command line arguments.
+     */
+    public static void main(final String[] args) {
+        SpringApplication.run(PrsApplication.class, args);
+    }
+
+}
