+++ coreservices/partsrelationshipservice/src/main/resources/application.yml
