+++ coreservices/partsrelationshipservice/src/test/java/net/catenax/prs/PrsApplicationTests.java
